test/requirements.txt
pytest==7.4.3
requests==2.31.0
pytest-mock==3.12.0
PyJWT==2.8.0
